applications/user-service/src/services/auth.py
"""Authentication utilities (JWT mock)"""
from jose import jwt
from datetime import datetime, timedelta
import logging

logger = logging.getLogger(__name__)

# Mock secret key (use environment variable in production)
SECRET_KEY = "helix-secret-key-change-in-production"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

def create_access_token(data: dict, expires_delta: timedelta = None) -> str:
    """Create JWT access token (mock implementation)"""
    to_encode = data.copy()
    
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    
    return encoded_jwt
